+++ fearless_simd/src/lib.rs
@@ -17,36 +17,29 @@
 //! These can be created in a SIMD context using the [`SimdFrom`] trait, or the
 //! [`from_slice`][SimdBase::from_slice] associated function.
 //!
-//! To create a function which SIMD and can be multiversioned, it will have a signature like:
+//! To call a function with the best available target features and get the associated `Simd`
+//! implementation, use the [`dispatch!()`] macro:
 //!
 //! ```rust
-//! use fearless_simd::{Simd, simd_dispatch};
+//! use fearless_simd::{Level, Simd, dispatch};
 //!
 //! #[inline(always)]
-//! fn sigmoid_impl<S: Simd>(simd: S, x: &[f32], out: &mut [f32]) { /* ... */ }
+//! fn sigmoid<S: Simd>(simd: S, x: &[f32], out: &mut [f32]) { /* ... */ }
 //!
-//! simd_dispatch!(fn sigmoid(level, x: &[f32], out: &mut [f32]) = sigmoid_impl);
+//! dispatch!(Level::new(), { #[inline(always)] |simd| sigmoid(simd, &[/*...*/], &mut [/*...*/]) });
 //! ```
 //!
 //! A few things to note:
 //!
-//! 1) This is generic over any `Simd` type.
-//! 2) The [`simd_dispatch`] macro is used to create a multi-versioned version of the given function.
-//! 3) The `_impl` suffix is used by convention to indicate the version of a function which will be dispatched to.
-//! 4) The `impl` function *must* be `#[inline(always)]`.
-//!    The performance of the SIMD implementation will be poor if that isn't the case. See [the section on inlining for details](#inlining)
+//! 1) `sigmoid` is generic over any `Simd` type.
+//! 2) The [`dispatch`] macro is used to invoke the given function with the target features associated with the supplied [`Level`].
+//! 3) The function or closure passed to [`dispatch!()`] should be `#[inline(always)]`.
+//!    The performance of the SIMD implementation may be poor if that isn't the case. See [the section on inlining for details](#inlining)
 //!
-//! The signature of the generated function will be:
-//!
-//! ```rust
-//! use fearless_simd::Level;
-//! fn sigmoid(level: Level, x: &[f32], out: &mut [f32]) { /* ... */ }
-//! ```
-//!
-//! The first parameter to this function is the [`Level`].
+//! The first parameter to [`dispatch!()`] is the [`Level`].
 //! If you are writing an application, you should create this once (using [`Level::new`]), and pass it to any function which wants to use SIMD.
 //! This type stores which instruction sets are available for the current process, which is used
-//! in the (generated) `sigmoid` function to dispatch to the most optimal variant of the function for this process.
+//! in the macro to dispatch to the most optimal variant of the supplied function for this process.
 //!
 //! # Inlining
 //!
@@ -169,7 +162,7 @@ pub enum Level {
     #[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
     Avx2(Avx2),
     // If new variants are added, make sure to handle them in `Level::dispatch`
-    // and `simd_dispatch`
+    // and `dispatch!()`
 }
 
 impl Level {
@@ -196,7 +189,7 @@ impl Level {
     /// you should construct the relevant variants yourself, using whatever
     /// way your specific chip supports accessing the current level.
     ///
-    /// This value will be passed to functions generated using [`simd_dispatch`].
+    /// This value should be passed to [`dispatch!()`].
     #[cfg(any(feature = "std", target_arch = "wasm32"))]
     #[must_use]
     pub fn new() -> Self {
@@ -318,7 +311,7 @@ impl Level {
 
     /// Dispatch `f` to a context where the target features which this `Level` proves are available are [enabled].
     ///
-    /// Most users of Fearless SIMD should prefer to use [`simd_dispatch`] to
+    /// Most users of Fearless SIMD should prefer to use [`dispatch!()`] to
     /// explicitly vectorize a function. That has a better developer experience
     /// than an implementation of `WithSimd`, and is less likely to miss a vectorization
     /// opportunity.